Finally fixed!
Only foreseeable issue is that on server stop the block will be changed
to air instead of originals. However, that is very minor. Now the chests
are unbreakable, and they disappear when the server is stopped.

Added bonus: the contents are still there on server start! (In case a
crash happens and players need their items)