I have now learned dependency injection!
To be honest the technique is fairly obvious. But now I know how to pass
instances of objects to each other and let them handle each other
correctly. There is a lot of bad code in here (being the first "major"
thing I ever wrote on my own) that will probably never be fixed because
it is not worth my time. But later commits should see better avoidance
of static and more emphasis on both readability and time complexity
analysis.